Tests/LibWeb/Text/expected/indexeddb-abort-upgrade-reverts-handle-names.txt

before rename: store=original_store index=original_index
after rename: store=renamed_store index=renamed_index
after abort: store=original_store index=original_index
